[S-275] Pinecone delete index vectors if namespace exists

+++ backend/app/services/vector_database/pinecone.py
@@ -52,7 +52,8 @@ class VectorDatabasePinecone(VectorDatabaseProvider):
         # Delete the index if it exists and create a new one. This way, every time
         # we upload vectors, we ensure that the index is clean and contains only the
         # latest vectors.
-        index.delete(delete_all=True, namespace=namespace)
+        if namespace in index.describe_index_stats().namespaces:
+            index.delete(delete_all=True, namespace=namespace)
 
         # Store the vectors.
         index.upsert(vectors=vectors, namespace=namespace)
